Document CRUD functionality
Lots of document CRUD updates. Most operations are in place and
functional now.

Two large issues this work uncovered.

1) Shared state between "urls." URLs imply a different stateful
resource, but when resources can be shared between multiple pages,
then the notion of mapping actions to URLs is less useful, and becomes
more complicated. For instance, the doc editor provides both a code
editor and a field editor, binding the data between the two gets
tricky as the two editors are different pages and different URLs. In
this case, we want to be able to share state between different child
pages, or more generally, we want the ability to share state between
subtabs and subpages.

2) When you have pages and subpages with nearly identical structure,
we still have a huge amount of duplicated boilerplate. For instance
the code and field editors use the same overall structure, but one
view is swapped out. We should have a way to remove the duplicated
boiler plate allowing each of those pages to be easily gernerated.
Alternatively, we could have subpages define what views they want to
swap out, and just extend from a parent page.

This commit has a handful of other random changes, like switching
pageUrl to just url with an optional context for switching on.